requirements.txt
# ============================================================
# Kyro — TRIAGE-AI  |  Python Dependencies
# ============================================================

# Flask
flask==3.1.0
python-dotenv==1.0.1

# Supabase
supabase==2.11.0

# Validation
marshmallow==3.23.1

# AI / ML
xgboost==2.1.3
scikit-learn==1.6.1
shap==0.46.0
numpy==1.26.4
pandas==2.2.3

# Production WSGI
gunicorn==23.0.0

# Logging (stdlib, no extra dep)
